# openapi.yaml
openapi: 3.0.1
info:
  title: Delilah Resto
  description: API Restaurant.
  contact:
    email: a-d-r-u@hotmail.com
  version: 1.0.0
...
servers:
  - url: 'https://localhost:3000'
  - url: 'http://localhost:3000'
...
paths:
  /users:
    post:
      tags:
        - User
      summary: Create user
      description: Creates a new User
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      security: []
      responses:
        '200':
          description: Your account has been created
        '400':
          description: Error validating data
        '409':
          description: 'Username already exists, please use another'
        default:
          $ref: '#/components/responses/Default'
    get:
      tags:
        - User
      summary: Lists users (Admin).
      description: >-
        Gets a list of all users in the DataBase with corresponding information.
        You should have admin credentials to access it.
      responses:
        '200':
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Default'
    put:
      tags:
        - User
      summary: Updates user information.
      description: >-
        Updates user information. You can update all the information except your password. You can perform this request only if you want to update your own information or if you have an admin profile. 
      requestBody:
        description: Information to update
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User was modified correctly
        '400':
          description: Unauthorized - Invalid Token or password
  
        '404':
          description: User not found
        '409':
          description: Username/email already exists, use another
        default:
          $ref: '#/components/responses/Default'
    delete:
      tags:
        - User
      summary: Deletes user account (Admin)
      responses:
        '200':
          description: User {username} was deleted
        '404':
          description: User not found
        default:
          $ref: '#/components/responses/Default'
  /users/login:
    get:
      tags:
        - User
      security: []
      summary: Log in user into the API.
      parameters:
        - name: username or email
          in: query
          required: true
          schema:
            type: string
        - name: passworD
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Invalid username/password supplied
        '401':
          description: Invalid request, user account disabled
        '422':
          description: mail/username and password are required
        default:
          $ref: '#/components/responses/Default'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: Camilo
        password:
          type: string
          example: camilo123
        email:
          type: string
          example: cam123@123.com
        delivery_address:
          type: string
          example: avenida 123
        full_name:
          type: string
          example: Camilo Ariza
        phone:
          type: integer
          example: 8565561
    UserInfo:
      type: array
      items:
        type: object
        properties:
          user_id:
            type: integer
            example: 1
          username:
            type: string
            example: Adriana
          full_name:
            type: string
            example: Adriana Ariza
          email:
            type: string
            example: a-d-r-u@hotmail.com
          phone:
            type: integer
            example: 5473214
          delivery_address:
            type: string
            example: calle123
          is_admin:
            type: boolean
            example: 1
          is_disabled:
            type: boolean
            example: 0
...
    Token:
      type: object
      properties:
        token:
          type: string
          example: >-
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6IkFkcmlhbmEiLCJ1c2VyX2lkIjoxLCJpc19hZG1pbiI6MSwiaWF0IjoxNjAyNTU1NzUwfQ.mVPDaGykUv0hjE8oayxE3QRErnh-QW2QhW32HS-y7hk
    Error:
      type: object
      properties:
        error:
          type: string
          example: 'An error has ocurred: <error>'
      required:
        - error
    Send:
      type: string
      example: Error message
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: The specified resource was not found
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Send'
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
...